requirements.txt
streamlit
biopython
google-genai
fpdf2
